requirements.txt
requests
schedule
numpy
scipy
python-dotenv
praw
instabot
Pillow